requirements.txt
streamlit
python-dotenv
google-generativeai
weaviate-client
pypdf
langchain
